Major refactor:
- `utils.py` split in three different classes:
  - `types.py` defines all the custom struct to be use.
  - `service.py` keep all the business logic.
  - `utils.py` manage all the format and message to be send.

- `cli.py` add the entry point for the sauron app, for the moment we
  only support telegram.
