+++ pyproject.toml
@@ -1,4 +1,4 @@
-# sauron-bot: telegram bot for monitoring telos block producer
+# sauron: bot for monitoring telos block producer
 # Copyright (C) 2024-present (in stewardship of openrepublic)
 [build-system]
 requires = ["hatchling"]
@@ -9,9 +9,9 @@ build-backend = "hatchling.build"
 ignore = []
 
 [project]
-name = "sauron-bot"
+name = "sauron"
 version = "0.1.0"
-description = "telegram bot for monitoring telos block producer"
+description = "bot for monitoring telos block producer"
 authors = [{ name = "Nelson Torres", email = "nelson.torres.a@proton.me" }]
 requires-python = ">=3.12, <3.13"
 license = "AGPL-3.0-or-later"
@@ -28,6 +28,7 @@ classifiers = [
 ]
 
 dependencies = [
+    'aiohttp==3.9.5, <4.0.0',
     'anyio >=3.7.1, <4.0.0',
     'click >=7.1.2, <8.0.0',
     'msgspec >=0.18.6, <0.19.0',
@@ -46,11 +47,14 @@ dev = [
     "pytest >=6.0.0, <7.0.0",
 ]
 
+[project.scripts]
+sauron = "sauron.cli:sauron"
+
 [tool.hatch.build.targets.sdist]
-include = ["sauron-bot"]
+include = ["sauron"]
 
 [tool.hatch.build.targets.wheel]
-include = ["sauron-bot"]
+include = ["sauron"]
 
 [tool.uv.sources]
 py-leap = { git = "https://github.com/guilledk/py-leap.git", tag = "v0.1a25" }
