+++ sauron/telegram.py
@@ -1,5 +1,4 @@
 import json
-import time
 import click
 import asyncio
 import msgspec
@@ -9,23 +8,21 @@ from leap.cleos import CLEOS
 from leap.protocol.ds import get_tapos_info 
 from telebot.async_telebot import AsyncTeleBot
 from telebot.types import CallbackQuery, Message
+from .utils import *
+from .service import *
 
 
-@click.command()
-@click.argument('filename', type=click.Path(exists=True))
-def telegram_bot(filename):
+def launch_telegram(filename):
 
-    utils = importlib.import_module('utils')
-    config = utils.get_config(filename)
+    config = get_config(filename)
 
     ntp_client = NTPClient()
-    bot = AsyncTeleBot(config.bot_token, exception_handler=utils.CustomExceptionHandler())
+    bot = AsyncTeleBot(config.bot_token, exception_handler=CustomExceptionHandler())
     cleos = CLEOS(endpoint=config.node_url)
-    cleos_local = CLEOS(endpoint=config.local_node_url)
 
     global system_status_cache
     global missed_bpr_cache
-    system_status_cache = utils.Cache()
+    system_status_cache = Cache()
     missed_bpr_cache = 0
 
     async def _async_main():
@@ -33,21 +30,27 @@ def telegram_bot(filename):
         async def refresh_status_cache(resource: str):
             while True:
                 global system_status_cache
+                import time
                 start_time = int(time.time())
-                system_status_cache.network = await asyncio.to_thread(utils.get_network_status)
+                system_status_cache.network = await asyncio.to_thread(get_network_status)
                 finished_time = int(time.time())
-                sleep_time = utils.sleep_delta(finished_time - start_time, resource)
+                sleep_time = sleep_delta(finished_time - start_time, resource)
                 await asyncio.sleep(sleep_time)
+
 
         async def send_notification():
             while True:
                 try:
                     global missed_bpr_cache
                     global system_status_cache
-                    system_status_cache.system = await utils.get_system_info()
-                    response, missed_bpr_cache = await utils.build_producer_status_message(
-                        cleos, cleos_local, ntp_client, system_status_cache , config.producer_name,
-                        config.users_alerted, missed_bpr_cache)
+                    system_status_cache.system = await get_system_info()
+                    response, missed_bpr_cache = await build_producer_status_message(
+                        cleos,
+                        ntp_client,
+                        system_status_cache,
+                        config,
+                        missed_bpr_cache
+                    )
                     await bot.send_message(config.chat_id, response, parse_mode='HTML')
                 except Exception as e:
                     print(f'An exception occurred: {e}')
@@ -83,6 +86,7 @@ def telegram_bot(filename):
                     ),
                     parse_mode='HTML')
 
+
         @bot.message_handler(commands=['u'])
         async def send_unregprod(message):
             ref_block_num, ref_block_prefix = get_tapos_info(
@@ -104,6 +108,7 @@ def telegram_bot(filename):
                         f"<i><u>tx_id:</u></i> <code>{res['transaction_id']}</code>"
                     ),
                     parse_mode='HTML')
+
 
         #@bot.message_handler(commands=['c'])
         async def request_claim_rewards(message):
@@ -128,34 +133,38 @@ def telegram_bot(filename):
                     ),
                     parse_mode='HTML')
 
+
         @bot.message_handler(commands=['schedule'])
         async def request_producers_schedule(message):
-            producers = cleos_local.get_schedule()['active']['producers']
-            schedule = utils.get_schedule_message([ producer['producer_name'] for producer in producers ], config.producer_name)
+            producers = await get_all_producers(config.node_url)
+            schedule = get_schedule_message([ producer['owner'] for producer in producers ], config.producer_name)
             await bot.reply_to(message=message, text=schedule, parse_mode='HTML')
+
 
         @bot.message_handler(commands=['s'])
         async def request_producer_status(message):
             global system_status_cache
             global missed_bpr_cache
-            system_status_cache.system = await utils.get_system_info()
-            response, missed_bpr_cache = await utils.build_producer_status_message(
-                    cleos, cleos_local, ntp_client, system_status_cache, config.producer_name,
-                    config.users_alerted, missed_bpr_cache)
+            system_status_cache.system = await get_system_info()
+            response, missed_bpr_cache = await build_producer_status_message(
+                cleos,
+                ntp_client,
+                system_status_cache,
+                config,
+                missed_bpr_cache)
             await bot.reply_to(message=message, text=response, parse_mode='HTML')
+
 
         @bot.message_handler(commands=['h'])
         async def request_help_message(message):
-            await bot.reply_to(message=message, text=utils.build_help_message(), parse_mode='HTML')
+            await bot.reply_to(message=message, text=build_help_message(), parse_mode='HTML')
 
-        utils.get_abi(cleos, config.abi_path)
+
+        get_abi(cleos, config.abi_path)
 
         asyncio.create_task(refresh_status_cache('network'))
         asyncio.create_task(send_notification())  
         await bot.infinity_polling()
-        
-    asyncio.run(_async_main())
 
-if __name__ == '__main__':
-    telegram_bot()
+    asyncio.run(_async_main())
